+++ e2e/mocks/config.service.ts
@@ -3,6 +3,6 @@ import { Injectable } from '@nestjs/common';
 @Injectable()
 export class ConfigService {
 	getHost() {
-		return 'localhost';
+		return '192.168.1.35';
 	}
 }