+++ lib/constants.ts
@@ -1,14 +1,17 @@
 export const RMQ_ROUTES_META: string = 'RMQ_ROUTES_META';
 export const RMQ_MESSAGE_META: string = 'RMQ_MESSAGE_META';
+export const RMQ_ROUTES_OPTIONS: string = 'RMQ_ROUTES_OPTIONS';
+export const RMQ_ROUTES_PATH: string = 'RMQ_ROUTES_PATH';
+export const RMQ_MESSAGE_INDEX: string = 'RMQ_MESSAGE_INDEX';
+export const RMQ_MODULE_OPTIONS: string = 'RMQ_MODULE_OPTIONS';
+
 export const DISCONNECT_EVENT: string = 'disconnect';
 export const CONNECT_EVENT: string = 'connect';
 export const DISCONNECT_MESSAGE: string = 'Disconnected from RMQ. Trying to reconnect';
 export const CONNECTED_MESSAGE: string = 'Successfully connected to RMQ';
 export const REPLY_QUEUE: string = 'amq.rabbitmq.reply-to';
 export const ERROR_NONE_RPC: string = 'This is none RPC queue. Use notify() method instead';
-// tslint:disable-next-line: quotemark
-export const ERROR_NO_ROUTE: string = "Requested service doesn't have RMQRoute with this path";
-export const ERROR_NO_ROUTE_FOR_CONTROLLER: string = 'No RMQRoute specified for controller';
+export const ERROR_NO_ROUTE: string = 'Requested service doesn\'t have RMQRoute with this path';
 export const ERROR_UNDEFINED_FROM_RPC: string = 'RPC method returned undefined';
 export const ERROR_TIMEOUT: string = 'Response timeout error';