+++ lib/decorators/rmq-message.decorator.ts
@@ -1,5 +1,4 @@
 import 'reflect-metadata';
-
 import { RMQ_MESSAGE_META } from '../constants';
 
 export function RMQMessage(target: any, propertyKey: string | symbol, parameterIndex: number) {