+++ lib/decorators/rmq-route.decorator.ts
@@ -1,14 +1,15 @@
-import { RMQ_ROUTES_META } from '../constants';
-import { IRouteMeta, IRouteOptions } from '../interfaces/queue-meta.interface';
-import { RMQService } from '../rmq.service';
+import {
+	RMQ_ROUTES_OPTIONS, RMQ_ROUTES_PATH,
+} from '../constants';
+import { IRouteOptions } from '../interfaces/queue-meta.interface';
+import { applyDecorators, SetMetadata } from '@nestjs/common';
 
-export const RMQRoute = (topic: string, options?: IRouteOptions) => {
-	return (target: any, methodName: string, descriptor: PropertyDescriptor) => {
-		let routes: IRouteMeta[] = Reflect.getMetadata(RMQ_ROUTES_META, RMQService);
-		if (!routes) {
-			routes = [];
-		}
-		routes.push({ topic, methodName, target, options });
-		Reflect.defineMetadata(RMQ_ROUTES_META, routes, RMQService);
-	};
+
+export const RMQRoute = (topic: string, options?: IRouteOptions): MethodDecorator => {
+		return applyDecorators(
+			SetMetadata(RMQ_ROUTES_OPTIONS, {
+				...options
+			}),
+			SetMetadata(RMQ_ROUTES_PATH, topic),
+		);
 };