+++ lib/interfaces/pipe-meta.interface.ts
@@ -1,7 +1,0 @@
-import { RMQPipeClass } from '../classes/rmq-pipe.class';
-
-export interface IPipeMeta {
-	pipe: typeof RMQPipeClass;
-	methodName: string;
-	target: any;
-}