+++ lib/interfaces/rmq-publish-options.interface.ts
@@ -1,4 +1,5 @@
 import { Options } from 'amqplib';
+
 export interface IPublishOptions extends Options.Publish {
 	timeout?: number;
 }