+++ lib/interfaces/rmq-router.interface.ts
@@ -1,4 +1,0 @@
-export interface IRMQRouter {
-	route: string;
-	propertyKey: any;
-}